A little update on CMake
It could compile well on Windows prior to win10.
(DX12 Render is disabled by default, To enable DX12, set DISABLE_D3D12
to 0. It is 1 by default. The best way to do it is to set it
automatically according to the OS version. I've no idea how to get that
information from CMake for now. So it is kind of a work around.)